+++ spec/spec_helper.rb
@@ -36,8 +36,10 @@ RSpec.configure do |config|
   end
   config.before(:each) do
     ::Que.clear!
-    expect(Que.execute('select * from que_scheduler_audit').count).to eq(0)
-    expect(Que.execute('select * from que_scheduler_audit_enqueued').count).to eq(0)
+    qsa = Que::Scheduler::VersionSupport.execute('select * from que_scheduler_audit')
+    expect(qsa.count).to eq(0)
+    qsae = Que::Scheduler::VersionSupport.execute('select * from que_scheduler_audit_enqueued')
+    expect(qsae.count).to eq(0)
   end
   config.before(:suite) do
     DbSupport.setup_db