+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+## Unreleased
+
+- Add facade to handle Que.execute differences between versions [#101](https://github.com/hlascelles/que-scheduler/pull/101)
+
 ## 3.2.4 (2019-07-14)
 
 - Add re-enqueue checks [#76](https://github.com/hlascelles/que-scheduler/pull/76)
 - Introduction of initial que 1.0 handling code [#95](https://github.com/hlascelles/que-scheduler/pull/95)
-- Add tests for Que 0.14 [#96](https://github.com/hlascelles/que-scheduler/pull/96)
+- Add tests for Que 0.14 [#97](https://github.com/hlascelles/que-scheduler/pull/97)
 
 ## 3.2.3 (2019-03-06)