+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 3.2.5 (2019-07-26)
 
 - Add facade to handle Que.execute differences between versions [#101](https://github.com/hlascelles/que-scheduler/pull/101)
 - Add facade to handle default queue differences between versions [#103](https://github.com/hlascelles/que-scheduler/pull/103)